requirements.txt
alembic==1.0.2
appdirs==1.4.4
APScheduler==3.6.3
asn1crypto==1.4.0
atomicwrites==1.2.1
attrs==20.3.0
blinker==1.4
boto3==1.16.58
botocore==1.19.58
bugsnag==4.0.2
certifi==2020.12.5
cffi==1.14.4
cfgv==3.2.0
chardet==4.0.0
Click==7.0
coverage==5.3.1
coveralls==3.0.0
cryptography==2.7
distlib==0.3.1
docopt==0.6.2
factory-boy==2.11.1
Faker==0.9.2
filelock==3.0.12
flasgger==0.9.5
Flask==1.0.2
Flask-API==1.0
Flask-Cors==3.0.6
Flask-Migrate==2.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.8.1
identify==1.5.10
idna==2.10
inflect==1.0.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
jsonschema==3.2.0
Mako==1.0.7
MarkupSafe==1.1.0
mistune==0.8.4
more-itertools==4.3.0
nodeenv==1.5.0
numpy==1.19.4
packaging==20.8
pluggy==0.13.1
pre-commit==2.9.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==1.6.4
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
pytest-cov==2.10.1
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2020.5
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
rollbar==0.15.1
s3transfer==0.3.4
six==1.11.0
SQLAlchemy==1.3.0
termcolor==1.1.0
text-unidecode==1.2
toml==0.10.2
tzlocal==2.1
urllib3==1.26.2
virtualenv==20.2.2
WebOb==1.8.6
Werkzeug==0.14.1
